backend/requirements.txt
# FastAPI and Server Dependencies
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
python-jose[cryptography]==3.3.0

# AI and LLM Dependencies
litellm==1.17.9
openai==1.3.7
anthropic==0.8.1

# Database and Caching
supabase==2.0.2
redis==5.0.1
asyncpg==0.29.0

# Data Providers and Market Data
polygon-api-client==1.12.4
websocket-client==1.6.4
aiofiles==23.2.1

# Data Processing
pandas==2.1.4
numpy==1.24.3
pydantic==2.5.1

# HTTP and Networking
httpx==0.25.2
aiohttp==3.9.1
requests==2.31.0

# Logging and Monitoring
structlog==23.2.0
sentry-sdk[fastapi]==1.38.0

# Security
python-dotenv==1.0.0
cryptography==41.0.8
bcrypt==4.1.2

# Testing (for development)
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0
httpx==0.25.2  # For testing API endpoints

# Development Tools
black==23.11.0
flake8==6.1.0
mypy==1.7.1